# Workr-backend/workr-swagger-api.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "Workr API",
    "version" : "2.1.0-oas3",
    "title" : "Workr API",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/appUrls" : {
      "get" : {
        "tags" : [ "General" ],
        "summary" : "check for new urls",
        "operationId" : "appUrls",
        "parameters" : [ {
          "in" : "query",
          "name" : "hash",
          "description" : "a hash that acts as an url version check",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/appurls"
                }
              }
            }
          },
          "301" : {
            "description" : "NOT OK"
          },
          "308" : {
            "description" : "Current installed version is not valid"
          },
          "400" : {
            "description" : "bad request"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/register" : {
      "post" : {
        "tags" : [ "General" ],
        "summary" : "Registers a new user as employee",
        "operationId" : "addEmployee",
        "description" : "adds an user to a company",
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "409" : {
            "description" : "You already have an account"
          },
          "500" : {
            "description" : "Unknown error occured"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "username", "password", "email", "name", "companycode" ],
                "properties" : {
                  "username" : {
                    "type" : "string"
                  },
                  "password" : {
                    "type" : "string"
                  },
                  "email" : {
                    "type" : "string"
                  },
                  "name" : {
                    "type" : "string"
                  },
                  "companycode" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "description" : "Returns the employee"
        }
      }
    },
...
    "/shift" : {
      "get" : {
        "tags" : [ "General" ],
        "summary" : "Loads the overview",
        "operationId" : "shift",
        "description" : "returns a list of all shifts",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Shift"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        }
      },
      "post" : {
        "tags" : [ "Employer" ],
        "summary" : "Create a shift",
        "operationId" : "createShift",
        "description" : "The employer can create a shift for a employee",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Shift"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "username", "isOpen", "date", "start", "end", "companycode" ],
                "properties" : {
                  "username" : {
                    "type" : "string"
                  },
                  "isOpen" : {
                    "type" : "boolean"
                  },
                  "date" : {
                    "type" : "string"
                  },
                  "start" : {
                    "type" : "string"
                  },
                  "end" : {
                    "type" : "string"
                  },
                  "companycode" : {
                    "type" : "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/shift/open" : {
      "get" : {
        "tags" : [ "General" ],
        "summary" : "show open shifts",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "operationId" : "openShifts",
        "description" : "returns a list of open shifts",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Shift"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        }
      }
    },
    "/shift/askfree" : {
      "put" : {
        "tags" : [ "General" ],
        "summary" : "sets the shift on available",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "operationId" : "askFree",
        "description" : "When asking free, the shift will be put on available till someone takes it over",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Shift"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "shiftid" ],
                "properties" : {
                  "shiftid" : {
                    "type" : "integer"
                  }
                }
              }
            }
          },
          "description" : "Returns the shift that has been set to free"
        }
      }
    },
    "/shift/cancelrequest" : {
      "put" : {
        "tags" : [ "General" ],
        "summary" : "cancels the shift request",
        "operationId" : "cancelRequest",
        "description" : "It will cancel the request that was put open",
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Shift"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Something went wrong"
          },
          "403" : {
            "$ref" : "#/components/responses/forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/usernotfound"
          }
        },
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "shiftid" ],
                "properties" : {
                  "shiftid" : {
                    "type" : "integer"
                  }
                }
              }
            }
          },
          "description" : "Returns the cancelled request"
        }
      }
    },
...
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/workr/workrAPI/1.0.0"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    },
    "schemas" : {
      "User" : {
        "type" : "object",
        "required" : [ "userId", "username", "name", "email", "picture", "phone" ],
        "properties" : {
          "userId" : {
            "type" : "integer",
            "example" : 12
          },
          "name" : {
            "type" : "string",
            "example" : "John Doe"
          },
          "username" : {
            "type" : "string",
            "example" : "JohnDoe"
          },
          "email" : {
            "type" : "string",
            "example" : "johndoe@gmail.com"
          },
          "picture" : {
            "type" : "string",
            "example" : "base64 string"
          },
          "phone" : {
            "type" : "string",
            "example" : 6.12345678E8
          }
        }
      },
      "Shift" : {
        "type" : "object",
        "required" : [ "User", "date", "start", "end", "shiftId", "open" ],
        "properties" : {
          "User" : {
            "$ref" : "#/components/schemas/User"
          },
          "date" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2019-11-11T00:00:00.000Z"
          },
          "start" : {
            "type" : "string",
            "format" : "time",
            "example" : "08:00 a.m."
          },
          "end" : {
            "type" : "string",
            "format" : "time",
            "example" : "11:00 a.m."
          },
          "shiftId" : {
            "type" : "integer",
            "example" : 13
          },
          "open" : {
            "type" : "boolean",
            "example" : true
          }
        }
      },
      "appurls" : {
        "type" : "array",
        "required" : [ "name", "url" ],
        "items" : {
          "properties" : {
            "name" : {
              "type" : "string",
              "example" : "profile"
            },
            "start" : {
              "type" : "string",
              "example" : "/user/profile"
            }
          }
        }
      }
    },
    "responses" : {
      "usernotfound" : {
        "description" : "User not found"
      },
      "forbidden" : {
        "description" : "You are not logged in."
      }
    }
  }
}